pipeline/requirements.txt
pytest
pylint
boto3
python-dotenv
psycopg2-binary
confluent_kafka